+++ src/CVMap.cpp
@@ -6,7 +6,6 @@ namespace StoermelderPackOne {
 namespace CVMap {
 
 static const int MAX_CHANNELS = 32;
-static const float UINIT = std::numeric_limits<float>::infinity();
 
 struct CVMapModule : CVMapModuleBase<MAX_CHANNELS> {
 	enum ParamIds {
@@ -28,19 +27,22 @@ struct CVMapModule : CVMapModuleBase<MAX_CHANNELS> {
 
 	/** [Stored to JSON] */
 	int panelTheme = 0;
-	/** [Stored to Json] */
+	/** [Stored to JSON] */
 	bool audioRate;
-	/** [Stored to Json] */
+	/** [Stored to JSON] */
 	bool locked;
 
 	dsp::ClockDivider processDivider;
 	dsp::ClockDivider lightDivider;
 
+	ScaledMapParam<float> mapParam[MAX_CHANNELS];
+
 	CVMapModule() {
 		panelTheme = pluginSettings.panelThemeDefault;
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
 		for (int i = 0; i < MAX_CHANNELS; i++) {
-			this->paramHandles[i].text = string::f("CV-MAP Ch%02d", i + 1);
+			paramHandles[i].text = string::f("CV-MAP Slot %02d", i + 1);
+			mapParam[i].setLimits(0.f, 1.f, std::numeric_limits<float>::infinity());
 		}
 		processDivider.setDivision(32);
 		lightDivider.setDivision(1024);
@@ -48,15 +50,24 @@ struct CVMapModule : CVMapModuleBase<MAX_CHANNELS> {
 	}
 
 	void onReset() override {
-		audioRate = true;
-		locked = false;
 		CVMapModuleBase<MAX_CHANNELS>::onReset();
+		audioRate = false;
+		locked = false;
+		for (size_t i = 0; i < MAX_CHANNELS; i++) {
+			mapParam[i].reset();
+		}
 	}
 
 	void process(const ProcessArgs& args) override {
 		if (audioRate || processDivider.process()) {
+			float deltaTime = args.sampleTime * (audioRate ? 1.f : float(processDivider.getDivision()));
+
 			// Step channels
 			for (int i = 0; i < mapLen; i++) {
+				ParamQuantity* paramQuantity = getParamQuantity(i);
+				if (paramQuantity == NULL) continue;
+				mapParam[i].setParamQuantity(paramQuantity);
+
 				if (i < 16) {
 					// Skip unused channels on INPUT1
 					if (inputs[POLY_INPUT1].getChannels() == i) {
@@ -70,24 +81,15 @@ struct CVMapModule : CVMapModuleBase<MAX_CHANNELS> {
 					}
 				}
 
-				ParamQuantity* paramQuantity = getParamQuantity(i);
-				if (paramQuantity == NULL) continue;
+				float t = (i < 16 ? inputs[POLY_INPUT1].getVoltage(i) : inputs[POLY_INPUT2].getVoltage(i - 16));
+				if (bipolarInput) t += 5.f;
+				t /= 10.f;
 
-				float v = i < 16 ? inputs[POLY_INPUT1].getVoltage(i) : inputs[POLY_INPUT2].getVoltage(i - 16);
-				if (bipolarInput)
-					v += 5.f;
-				v = rescale(v, 0.f, 10.f, 0.f, 1.f);
+				// Set a new value for the mapped parameter
+				mapParam[i].setValue(t);
 
-				// If lastValue is unitialized set it to its current value, only executed once
-				if (lastValue[i] == UINIT) {
-					lastValue[i] = v;
-				}
-
-				if (lockParameterChanges || lastValue[i] != v) {
-					// Set ParamQuantity
-					paramQuantity->setScaledValue(v);
-					lastValue[i] = v;
-				}
+				// Apply value on the mapped parameter (respecting slew and scale)
+				mapParam[i].process(deltaTime, lockParameterChanges);
 			}
 		}
 
@@ -114,6 +116,12 @@ struct CVMapModule : CVMapModuleBase<MAX_CHANNELS> {
 		return rootJ;
 	}
 
+	void dataToJsonMap(json_t* mapJ, int index) override {
+		json_object_set_new(mapJ, "slew", json_real(mapParam[index].getSlew()));
+		json_object_set_new(mapJ, "min", json_real(mapParam[index].getMin()));
+		json_object_set_new(mapJ, "max", json_real(mapParam[index].getMax()));
+	}
+
 	void dataFromJson(json_t* rootJ) override {
 		CVMapModuleBase<MAX_CHANNELS>::dataFromJson(rootJ);
 		panelTheme = json_integer_value(json_object_get(rootJ, "panelTheme"));
@@ -122,6 +130,30 @@ struct CVMapModule : CVMapModuleBase<MAX_CHANNELS> {
 		if (audioRateJ) audioRate = json_boolean_value(audioRateJ);
 		json_t* lockedJ = json_object_get(rootJ, "locked");
 		if (lockedJ) locked = json_boolean_value(lockedJ);
+	}
+
+	void dataFromJsonMap(json_t* mapJ, int index) override {
+		json_t* slewJ = json_object_get(mapJ, "slew");
+		json_t* minJ = json_object_get(mapJ, "min");
+		json_t* maxJ = json_object_get(mapJ, "max");
+		if (slewJ) mapParam[index].setSlew(json_real_value(slewJ));
+		if (minJ) mapParam[index].setMin(json_real_value(minJ));
+		if (maxJ) mapParam[index].setMax(json_real_value(maxJ));
+	}
+};
+
+
+struct CvMapChoice : MapModuleChoice<MAX_CHANNELS, CVMapModule> {
+	void appendContextMenu(Menu* menu) override {
+		menu->addChild(new MenuSeparator());
+		menu->addChild(new MapSlewSlider<>(&module->mapParam[id]));
+		menu->addChild(new MenuSeparator());
+		menu->addChild(construct<MenuLabel>(&MenuLabel::text, "Scaling"));
+		menu->addChild(construct<MapScalingInputLabel<>>(&MenuLabel::text, "Input", &MapScalingInputLabel<>::p, &module->mapParam[id]));
+		menu->addChild(construct<MapScalingOutputLabel<>>(&MenuLabel::text, "Parameter range", &MapScalingOutputLabel<>::p, &module->mapParam[id]));
+		menu->addChild(new MapMinSlider<>(&module->mapParam[id]));
+		menu->addChild(new MapMaxSlider<>(&module->mapParam[id]));
+		menu->addChild(construct<MapPresetMenuItem<>>(&MenuItem::text, "Presets", &MapPresetMenuItem<>::p, &module->mapParam[id]));
 	}
 };
 
@@ -147,7 +179,7 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 		w1->setModule(module, CVMapModule::CHANNEL_LIGHTS2);
 		addChild(w1);
 
-		typedef MapModuleDisplay<MAX_CHANNELS, CVMapModule> TMapDisplay;
+		typedef MapModuleDisplay<MAX_CHANNELS, CVMapModule, CvMapChoice> TMapDisplay;
 		TMapDisplay* mapWidget = createWidget<TMapDisplay>(Vec(10.6f, 81.5f));
 		mapWidget->box.size = Vec(128.9f, 261.7f);
 		mapWidget->setModule(module);
@@ -162,11 +194,9 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 
 		struct LockItem : MenuItem {
 			CVMapModule* module;
-
 			void onAction(const event::Action& e) override {
 				module->lockParameterChanges ^= true;
 			}
-
 			void step() override {
 				rightText = module->lockParameterChanges ? "Locked" : "Unlocked";
 				MenuItem::step();
@@ -175,11 +205,9 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 
 		struct UniBiItem : MenuItem {
 			CVMapModule* module;
-
 			void onAction(const event::Action& e) override {
 				module->bipolarInput ^= true;
 			}
-
 			void step() override {
 				rightText = module->bipolarInput ? "-5V..5V" : "0V..10V";
 				MenuItem::step();
@@ -188,11 +216,9 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 
 		struct AudioRateItem : MenuItem {
 			CVMapModule* module;
-
 			void onAction(const event::Action& e) override {
 				module->audioRate ^= true;
 			}
-
 			void step() override {
 				rightText = module->audioRate ? "✔" : "";
 				MenuItem::step();
@@ -201,11 +227,9 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 
 		struct TextScrollItem : MenuItem {
 			CVMapModule* module;
-
 			void onAction(const event::Action& e) override {
 				module->textScrolling ^= true;
 			}
-
 			void step() override {
 				rightText = module->textScrolling ? "✔" : "";
 				MenuItem::step();
@@ -214,11 +238,9 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 
 		struct MappingIndicatorHiddenItem : MenuItem {
 			CVMapModule* module;
-
 			void onAction(const event::Action& e) override {
 				module->mappingIndicatorHidden ^= true;
 			}
-
 			void step() override {
 				rightText = module->mappingIndicatorHidden ? "✔" : "";
 				MenuItem::step();
@@ -227,11 +249,9 @@ struct CVMapWidget : ThemedModuleWidget<CVMapModule> {
 
 		struct LockedItem : MenuItem {
 			CVMapModule* module;
-
 			void onAction(const event::Action& e) override {
 				module->locked ^= true;
 			}
-
 			void step() override {
 				rightText = module->locked ? "✔" : "";
 				MenuItem::step();