+++ CHANGELOG.md
@@ -31,6 +31,7 @@
     - Added an option for auto-loading the last active preset
     - Fixed broken "Autoload first preset" (#29)
 - Module [CV-MAP](./docs/CVMap.md)
+    - Added context menu sliders for slew and scaling and transforming the input and parameter-range (#243)
     - Fixed wrong slot behavior when toggling input to 0V (#221)
 - Module [GLUE](./docs/Glue.md)
     - Added option to consolidate all GLUE modules into the current one